Revised to allow for multiple compile states!
Recent changes to TCC allow for multiple compile states, finally!
However, it is not possible to create a second TCCState before
relocating the first. As such, I had to refactor the storage of
state information and push all of the manipulations of the TCCState
object into the single compile function.