Got TCC::Callable working!
I was confused about passing doubles and passing pointers. The
doubles confusion was a red herring: I was printing garbage in
the diganostic printf, and I was somehow invoking pow() in a way
that gave me garbage, too.

The pointers are presently resolved, but it's not clear to me if
they will always work this way or not. The problem was that I was
returning the pointer value via the XS routine PTR2IV, but it turns
out that you can't pack that as a 'p' or 'P' pointer. Precisely what
those arguments expect is not clear to me. But, if I pack the return
value as a Perl 'j', which is Perl's internal representation of IV,
everything Just Works. At least for now. :-)

It is now time to write docs and tests. :-)