Added #define _ANSI_SOURCE in StretchyBuffer
I encountered a weird issue with StretchyBuffer on my Mac in which it threw a
large number of compile-time errors. The somewhat hackish solution from the tcc
mailing list is to #define _ANSI_SOURCE. Since StretchyBuffer is the only
culprit for now, I have simply added that line to the package. Fixed for now!